requirements.txt
google-auth>=2.29.0
google-genai>=1.30.0
opencv-python>=4.9.0
pdf2image>=1.17.0
Pillow>=10.3.0
PyMuPDF>=1.24.0
pytesseract>=0.3.10
PyYAML>=6.0.1
tqdm>=4.66.0
